requirements.txt
pyTelegramBotAPI==4.14.1
openrouter==0.1.5
python-dotenv==1.0.1